test(e2e/tron): refactor account tests for speed
